Refactor Fiber Prometheus Middleware
- [+] refactor(middleware): remove NewPrometheusMiddleware function from utils.go
- [+] feat(k8s): add k8s package with metrics subpackage for Kubernetes-related functionality
- [+] feat(k8s/metrics): add NewPrometheusMiddleware function for creating Prometheus middleware
- [+] feat(server): add MountPrometheusMiddlewareHandler function for mounting Prometheus middleware
- [+] feat(server): add RegisterRoutesPrometheus method to FiberServer for registering Prometheus middleware
- [+] docs(k8s): add package documentation for k8s package and its subpackages
